Move advanced VM settings to a separate section to avoid layout weirdness
